Refactor into modular pages
