+++ pages/history.py
@@ -1,0 +1,29 @@
+import streamlit as st
+from logic import (
+    get_behavior_history,
+    get_behavior_history_by_filters,
+)
+from ui import select_filters, create_history_line_chart
+
+
+def run(df):
+    """Render the behavior history page."""
+    filter_option, sel_animal, sel_sex, sel_groups = select_filters(df, key_prefix="history_")
+    behaviors = df["Unified Behavior"].unique()
+    selected_behavior = st.selectbox("Select Behavior", behaviors, key="history_behavior")
+
+    if filter_option == "By Individual" and sel_animal:
+        df_line = get_behavior_history(df, sel_animal, selected_behavior)
+        title = f"{selected_behavior} over time for {sel_animal}"
+    else:
+        df_line = get_behavior_history_by_filters(
+            df,
+            sexes=sel_sex,
+            groups=sel_groups,
+            behavior=selected_behavior,
+        )
+        sex_text = ", ".join(sel_sex) if sel_sex else "All Sexes"
+        group_text = ", ".join(sel_groups) if sel_groups else "All Social Groups"
+        title = f"{selected_behavior} over time | Sex: {sex_text} | Group: {group_text}"
+
+    create_history_line_chart(df_line, title)
